+++ changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.13.6] - 2026-9-4 API Investigation
+
+### Bug Fixes
+- fixed string bug on /record where base url returned no url
+- fixed title on /schedule saying "Not Scheduled" when all players confirmed in LFG
+
 ## [1.13.5] - 2026-9-1 September Forever
 
 ### New Features